+++ packages/houdini-svelte/src/plugin/transforms/index.ts
@@ -5,7 +5,8 @@ import * as recast from 'recast'
 import type { SourceMapInput } from 'rollup'
 
 import { parseSvelte } from '../extract'
-import { plugin_config, type Framework } from '../kit'
+import { type Framework } from '../kit'
+import { plugin_config } from '../config'
 import query from './componentQuery'
 import kit from './kit'
 import tags from './tags'
