+++ packages/plugin-svelte-global-stores/src/plugin/index.ts
@@ -1,7 +1,7 @@
 import { plugin, type PluginHooks } from 'houdini'
 import { HoudiniError, path } from 'houdini'
 
-import { store_name } from '../../../houdini-svelte/src/plugin/kit'
+import { store_name } from '../../../houdini-svelte/src/plugin/storeConfig'
 import generate from './codegen'
 import { global_stores_directory, global_store_name } from './kit'
